requirements.txt
absl-py==0.11.0
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
black==20.8b1
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
confuse==1.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
gitpython==3.1.13
google-auth-oauthlib==0.4.2
google-auth==1.26.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
imagehash==4.2.0
ipykernel==5.4.3
ipython-genutils==0.2.0
ipython==7.20.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0rc3
markdown==3.3.3
markupsafe==1.1.1
matplotlib==3.3.4
missingno==0.4.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.2
nbconvert==6.0.7;
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas-profiling==2.10.1
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
phik==0.11.0
pickleshare==0.7.5
pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydeck==0.6.0
pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywavelets==1.1.1
pywin32==300
pywinpty==0.5.
pyyaml==5.4.1
pyzmq==22.0.2
regex==2020.11.13
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7
scipy==1.6.0
seaborn==0.11.1
send2trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.76.0
tangled-up-in-unicode==0.0.6
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1
tensorflow-estimator==2.4.0
tensorflow==2.4.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
visions==0.6.0
watchdog==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
